SQUASH: POC
