+++ q2_types/distance_matrix/__init__.py
@@ -1,0 +1,16 @@
+# ----------------------------------------------------------------------------
+# Copyright (c) 2016--, QIIME development team.
+#
+# Distributed under the terms of the Modified BSD License.
+#
+# The full license is in the file LICENSE, distributed with this software.
+# ----------------------------------------------------------------------------
+
+import importlib
+
+from ._format import LSMatFormat, DistanceMatrixDirectoryFormat
+from ._type import DistanceMatrix
+
+__all__ = ['LSMatFormat', 'DistanceMatrixDirectoryFormat', 'DistanceMatrix']
+
+importlib.import_module('q2_types.distance_matrix._transformer')
